requirements.txt
selenium==3.141.0
retrying==1.3.3
numpy==1.14.3
tqdm==4.62.0
pandas==0.23.0
pymysql==1.0.2
PyYAML==5.4.1
